Refactor cli into 3 clients
