+++ cmd/cli/cmd/componentShow.go
@@ -6,14 +6,8 @@
 package cmd
 
 import (
-	"encoding/json"
-	"fmt"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-	"github.com/Azure/radius/cmd/cli/utils"
 	"github.com/Azure/radius/pkg/rad"
-	"github.com/Azure/radius/pkg/radclient"
+	"github.com/Azure/radius/pkg/rad/environments"
 	"github.com/spf13/cobra"
 )
 
@@ -45,24 +39,10 @@ func showComponent(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	azcred, err := azidentity.NewDefaultAzureCredential(nil)
+	client, err := environments.CreateManagementClient(env)
 	if err != nil {
-		return fmt.Errorf("failed to obtain a Azure credentials: %w", err)
-	}
-	con := armcore.NewDefaultConnection(azcred, nil)
-	componentClient := radclient.NewComponentClient(con, env.SubscriptionID)
-
-	response, err := componentClient.Get(cmd.Context(), env.ResourceGroup, applicationName, componentName, nil)
-	if err != nil {
-		return utils.UnwrapErrorFromRawResponse(err)
+		return err
 	}
 
-	componentResource := *response.ComponentResource
-	componentDetails, err := json.MarshalIndent(componentResource, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal component response as JSON %w", err)
-	}
-	fmt.Println(string(componentDetails))
-
-	return err
+	return client.ShowComponent(cmd.Context(), applicationName, componentName)
 }
