+++ cmd/cli/cmd/deploymentDelete.go
@@ -8,12 +8,9 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-	"github.com/Azure/radius/cmd/cli/utils"
 	"github.com/Azure/radius/pkg/rad"
+	"github.com/Azure/radius/pkg/rad/environments"
 	"github.com/Azure/radius/pkg/rad/prompt"
-	"github.com/Azure/radius/pkg/radclient"
 	"github.com/spf13/cobra"
 )
 
@@ -46,14 +43,14 @@ func deleteDeployment(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	depName, err := rad.RequireDeployment(cmd, args)
+	deploymentName, err := rad.RequireDeployment(cmd, args)
 	if err != nil {
 		return err
 	}
 
 	// Prompt user to confirm deletion
 	if !yes {
-		confirmed, err := prompt.Confirm(fmt.Sprintf("Are you sure you want to delete '%v' from '%v' [y/n]?", depName, env.Name))
+		confirmed, err := prompt.Confirm(fmt.Sprintf("Are you sure you want to delete '%v' from '%v' [y/n]?", deploymentName, env.GetName()))
 		if err != nil {
 			return err
 		}
@@ -62,25 +59,10 @@ func deleteDeployment(cmd *cobra.Command, args []string) error {
 		}
 	}
 
-	azcred, err := azidentity.NewDefaultAzureCredential(nil)
+	client, err := environments.CreateManagementClient(env)
 	if err != nil {
-		return fmt.Errorf("failed to obtain Azure credential: %w", err)
+		return err
 	}
 
-	con := armcore.NewDefaultConnection(azcred, nil)
-
-	dc := radclient.NewDeploymentClient(con, env.SubscriptionID)
-	poller, err := dc.BeginDelete(cmd.Context(), env.ResourceGroup, applicationName, depName, nil)
-	if err != nil {
-		return utils.UnwrapErrorFromRawResponse(err)
-	}
-
-	_, err = poller.PollUntilDone(cmd.Context(), radclient.PollInterval)
-	if err != nil {
-		return utils.UnwrapErrorFromRawResponse(err)
-	}
-
-	fmt.Printf("Deployment '%s' deleted.\n", depName)
-
-	return err
+	return client.DeleteDeployment(cmd.Context(), applicationName, deploymentName)
 }
