+++ cmd/cli/cmd/deploymentList.go
@@ -6,17 +6,8 @@
 package cmd
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"net/http"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-	"github.com/Azure/radius/cmd/cli/utils"
 	"github.com/Azure/radius/pkg/rad"
-	"github.com/Azure/radius/pkg/radclient"
+	"github.com/Azure/radius/pkg/rad/environments"
 	"github.com/spf13/cobra"
 )
 
@@ -43,31 +34,10 @@ func listDeployments(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	azcred, err := azidentity.NewDefaultAzureCredential(nil)
+	client, err := environments.CreateManagementClient(env)
 	if err != nil {
-		return fmt.Errorf("failed to obtain Azure credentials: %w", err)
-	}
-	con := armcore.NewDefaultConnection(azcred, nil)
-	dc := radclient.NewDeploymentClient(con, env.SubscriptionID)
-
-	response, err := dc.ListByApplication(cmd.Context(), env.ResourceGroup, applicationName, nil)
-	if err != nil {
-		var httpresp azcore.HTTPResponse
-		if ok := errors.As(err, &httpresp); ok && httpresp.RawResponse().StatusCode == http.StatusNotFound {
-			errorMessage := fmt.Sprintf("application '%s' was not found in the resource group '%s'", applicationName, env.ResourceGroup)
-			return radclient.NewRadiusError("ResourceNotFound", errorMessage)
-		}
-
-		return utils.UnwrapErrorFromRawResponse(err)
+		return err
 	}
 
-	deploymentsList := *response.DeploymentList
-	deployments, err := json.MarshalIndent(deploymentsList, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal deployment response as JSON %w", err)
-	}
-
-	fmt.Println(string(deployments))
-
-	return err
+	return client.ListDeployments(cmd.Context(), applicationName)
 }
