+++ cmd/cli/cmd/deploymentShow.go
@@ -6,18 +6,8 @@
 package cmd
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"net/http"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/armcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-
-	"github.com/Azure/radius/cmd/cli/utils"
 	"github.com/Azure/radius/pkg/rad"
-	"github.com/Azure/radius/pkg/radclient"
+	"github.com/Azure/radius/pkg/rad/environments"
 	"github.com/spf13/cobra"
 )
 
@@ -44,36 +34,15 @@ func showDeployment(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	depName, err := rad.RequireDeployment(cmd, args)
+	deploymentName, err := rad.RequireDeployment(cmd, args)
 	if err != nil {
 		return err
 	}
 
-	azcred, err := azidentity.NewDefaultAzureCredential(nil)
+	client, err := environments.CreateManagementClient(env)
 	if err != nil {
-		return fmt.Errorf("failed to obtain Azure credentials: %w", err)
-	}
-	con := armcore.NewDefaultConnection(azcred, nil)
-	dc := radclient.NewDeploymentClient(con, env.SubscriptionID)
-
-	response, err := dc.Get(cmd.Context(), env.ResourceGroup, applicationName, depName, nil)
-	if err != nil {
-		var httpresp azcore.HTTPResponse
-		if ok := errors.As(err, &httpresp); ok && httpresp.RawResponse().StatusCode == http.StatusNotFound {
-			errorMessage := fmt.Sprintf("deployment '%s' for application '%s' and resource group '%s' was not found", depName, applicationName, env.ResourceGroup)
-			return radclient.NewRadiusError("ResourceNotFound", errorMessage)
-		}
-
-		return utils.UnwrapErrorFromRawResponse(err)
+		return err
 	}
 
-	deploymentResource := *response.DeploymentResource
-	deploymentDetails, err := json.MarshalIndent(deploymentResource, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal deployment response as JSON %w", err)
-	}
-
-	fmt.Println(string(deploymentDetails))
-
-	return err
+	return client.ShowDeployment(cmd.Context(), deploymentName, applicationName)
 }
