+++ cmd/cli/cmd/root.go
@@ -8,16 +8,11 @@ package cmd
 import (
 	"fmt"
 	"os"
-	"path"
 
+	"github.com/Azure/radius/pkg/rad"
 	"github.com/Azure/radius/pkg/version"
 	"github.com/spf13/cobra"
-
-	homedir "github.com/mitchellh/go-homedir"
-	"github.com/spf13/viper"
 )
-
-var cfgFile string
 
 // RootCmd is the root command of the rad CLI. This is exported so we can generate docs for it.
 var RootCmd = &cobra.Command{
@@ -46,55 +41,9 @@ func init() {
 	RootCmd.SetVersionTemplate(template)
 
 	RootCmd.Flags().BoolP("version", "v", false, "version for radius")
-	RootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.rad/config.yaml)")
+	RootCmd.PersistentFlags().StringVar(&rad.CfgFile, "config", "", "config file (default is $HOME/.rad/config.yaml)")
 }
 
-// initConfig reads in config file and ENV variables if set.
 func initConfig() {
-	if cfgFile != "" {
-		// Use config file from the flag.
-		viper.SetConfigFile(cfgFile)
-	} else {
-		// Find home directory.
-		home, err := homedir.Dir()
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
-
-		rad := path.Join(home, ".rad")
-		viper.AddConfigPath(rad)
-		viper.SetConfigName("config")
-	}
-
-	// If a config file is found, read it in.
-	err := viper.ReadInConfig()
-	if _, ok := err.(viper.ConfigFileNotFoundError); ok {
-		home, err := homedir.Dir()
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
-
-		// Set the default config file so we can write to it if desired
-		cfgFile = path.Join(home, ".rad", "config.yaml")
-	} else if err == nil {
-		cfgFile = viper.ConfigFileUsed()
-	}
-}
-
-func saveConfig() error {
-	dir := path.Dir(cfgFile)
-	_, err := os.Stat(dir)
-	if os.IsNotExist(err) {
-		_ = os.MkdirAll(dir, os.ModeDir|0755)
-	}
-
-	err = viper.WriteConfigAs(cfgFile)
-	if err != nil {
-		return err
-	}
-
-	fmt.Printf("Successfully wrote configuration to %v\n", cfgFile)
-	return nil
+	rad.InitConfig()
 }
