+++ pkg/rad/clients/clients.go
@@ -30,8 +30,20 @@ type ExposeOptions struct {
 type LogsOptions struct {
 	Application string
 	Component   string
+	Follow      bool
+	Container   string
 }
 
 // ManagementClient is used to interface with management features like listing applications and components.
 type ManagementClient interface {
+	ListApplications(ctx context.Context) error
+	ShowApplication(ctx context.Context, applicationName string) error
+	DeleteApplication(ctx context.Context, applicationName string) error
+
+	ListComponents(ctx context.Context, applicationName string) error
+	ShowComponent(ctx context.Context, applicationName string, componentName string) error
+
+	ListDeployments(ctx context.Context, applicationName string) error
+	ShowDeployment(ctx context.Context, deploymentName string, applicationName string) error
+	DeleteDeployment(ctx context.Context, deploymentName string, applicationName string) error
 }
